Add settings view
A user can and is required to enter the necessary information
in the settings view. That information is stored in the localStorage and loaded at the start of
the app.

Next, I need to implement functions for validating and parsing the prefixes.
After that, I need to implement the actual graph drawing logic.