Add service module for loading triples and simple graph component
The service module loads the triples from a given endpoint, however it appears this isn't the intended behaviour.
Instead, I'll likely have to directly request the triples from the triplestor using stardogjs.
The graph component displays the data, however the literal bug seems to still be there.
URIs are correctly replaced by prefixes.

Next, I need to add proper loading of the triples and subsequent processing and storing using graphy, for instance.
I also need to add the colouring, a zooming feature and a bounding box force, the info pop-up upon clicking a node
(containing the type and other information), shortening of text literals and the expanding behaviour
when the node is clicked and the separation of identical literal value nodes.