Add iteration 1 requirements
The triples are now all loaded from the given endpoint.
The triples are visualized as a zoomable and draggable graph.
The number of visualizized triples can be limited for performance improvements.
Clicking on a node will reveal its type and value.

I need to refactor the loadGraph function to clean it up, reduce its cognitive complexity,
make it more customizable and by extension also the graph component.
Big graphs are hard to see through both when zoomed in and out.
Big graphs act strangely when zoomed in by a lot.
The adjustments should be accessible more easily (from the graph view).