Revert predicates being nodes, Add default prefixes, reset buttion, arrows to edges, fix bugs and more
Predicates are no longer nodes. They are now links again with text showing their content.

Default prefixes are now fetched at the start of the app.

A button for resetting the simulation and thereby making fixed nodes dynamic again  has been added.

Markers showing the direction of edges have been added to the graph. They are only rendered for highlighted nodes.

The bug occuring in the settings view causing crashes when the window was resized has now been fixed.
The "bug" causing the loading symbol not to be shown when loading the graph has been fixed as well.

A disclaimer has been added to the prefix input field.

A rectangular background has been added to the edge text allowing for better vision.

The highlighting no longer modifies the stroke-width of an edge.
Instead, the stroke-width has been increased in general to improve visibility
and avoid problems with the arrow markers.