Add text toggle, force database in endpoint url, fix resize bug
The text toggling for both nodes and links has been added and can
now be configured in the settings view.

Supplying the database in the endpoint url is now mandatory since
the qanary database is no longer the default one.

The resizing bug which would cause the graph to be drawn twice
has now been fixed.