Add collapsing and expanding of nodes, rename custom attrs
The nodes can be collapsed based on a partial graphs algorithm
which gets all partial graphs (which would otherwise be connected
through the selected node) and hides all but the biggest and
the loose graphs/nodes. I've gone through a few other methods
such as dynamically determining a root node or having collapsing
be the inverse operation of expansion, but both have issues
which I may describe in more detail in a meeting.
The expansion is the inverse operation of the collapsion,
i. e. it adds the hidden partial graphs back in.
The collapse *may* be improved by either hiding both or neither equally large
partial graphs (e. g. collapsing the middle node of a three node graph
where the collapsed node is the only connection).

Further, the custom attributes of nodes and links have been renamed to
include underscores to indicate their status as custom attributes and
to avoid compatibility issues with future versions.