Add initial collapsing of nodes, improve collapsing effect
The initial collapsing hides nodes until the specified maximum
number of nodes is satisfied. The collapsing now hides all partial
graph if there are only equally large ones and maintains all equally
sized partial graphs if there is a smaller one.

It has to be noted somewhere that collapsing does not necessarily
recreate the initially collapsed graph and that collapsing and expanding
are not mutually inverse actions.